requirements.txt
# This file was autogenerated by uv via the following command:
#    uv export --format requirements-txt --no-hashes --output-file requirements.txt
aiofiles==24.1.0
    # via gradio
annotated-doc==0.0.4
    # via fastapi
annotated-types==0.7.0
    # via pydantic
anyio==4.12.1
    # via
    #   gradio
    #   httpx
    #   mcp
    #   openai
    #   sse-starlette
    #   starlette
attrs==25.4.0
    # via
    #   jsonschema
    #   referencing
audioop-lts==0.2.2
    # via gradio
brotli==1.2.0
    # via gradio
cachetools==6.2.4
    # via pyiceberg
certifi==2026.1.4
    # via
    #   httpcore
    #   httpx
    #   requests
cffi==2.0.0 ; platform_python_implementation != 'PyPy'
    # via cryptography
charset-normalizer==3.4.4
    # via requests
click==8.3.1
    # via
    #   pyiceberg
    #   typer
    #   typer-slim
    #   uvicorn
colorama==0.4.6
    # via
    #   click
    #   griffe
    #   pytest
    #   tqdm
coverage==7.13.1
    # via pytest-cov
cryptography==46.0.3
    # via pyjwt
deprecation==2.1.0
    # via
    #   postgrest
    #   storage3
distro==1.9.0
    # via openai
fastapi==0.128.0
    # via gradio
ffmpy==1.0.0
    # via gradio
filelock==3.20.3
    # via huggingface-hub
fsspec==2026.1.0
    # via
    #   gradio-client
    #   huggingface-hub
    #   pyiceberg
gradio==6.3.0
    # via ecommerce-support
gradio-client==2.0.3
    # via gradio
griffe==1.15.0
    # via openai-agents
groovy==0.1.2
    # via gradio
h11==0.16.0
    # via
    #   httpcore
    #   uvicorn
h2==4.3.0
    # via httpx
hf-xet==1.2.0 ; platform_machine == 'AMD64' or platform_machine == 'aarch64' or platform_machine == 'amd64' or platform_machine == 'arm64' or platform_machine == 'x86_64'
    # via huggingface-hub
hpack==4.1.0
    # via h2
httpcore==1.0.9
    # via httpx
httpx==0.28.1
    # via
    #   gradio
    #   gradio-client
    #   huggingface-hub
    #   mcp
    #   openai
    #   postgrest
    #   safehttpx
    #   storage3
    #   supabase
    #   supabase-auth
    #   supabase-functions
httpx-sse==0.4.3
    # via mcp
huggingface-hub==1.3.2
    # via
    #   gradio
    #   gradio-client
hyperframe==6.1.0
    # via h2
idna==3.11
    # via
    #   anyio
    #   httpx
    #   requests
    #   yarl
iniconfig==2.3.0
    # via pytest
jinja2==3.1.6
    # via gradio
jiter==0.12.0
    # via openai
jsonschema==4.26.0
    # via mcp
jsonschema-specifications==2025.9.1
    # via jsonschema
markdown-it-py==4.0.0
    # via rich
markupsafe==3.0.3
    # via
    #   gradio
    #   jinja2
mcp==1.25.0
    # via openai-agents
mdurl==0.1.2
    # via markdown-it-py
mmh3==5.2.0
    # via pyiceberg
multidict==6.7.0
    # via yarl
numpy==2.4.1
    # via
    #   gradio
    #   pandas
openai==2.15.0
    # via openai-agents
openai-agents==0.6.7
    # via ecommerce-support
orjson==3.11.5
    # via gradio
packaging==25.0
    # via
    #   deprecation
    #   gradio
    #   gradio-client
    #   huggingface-hub
    #   pytest
pandas==2.3.3
    # via gradio
pillow==12.1.0
    # via gradio
pluggy==1.6.0
    # via
    #   pytest
    #   pytest-cov
postgrest==2.27.2
    # via supabase
propcache==0.4.1
    # via yarl
pycparser==2.23 ; implementation_name != 'PyPy' and platform_python_implementation != 'PyPy'
    # via cffi
pydantic==2.12.5
    # via
    #   fastapi
    #   gradio
    #   mcp
    #   openai
    #   openai-agents
    #   postgrest
    #   pydantic-settings
    #   pyiceberg
    #   realtime
    #   storage3
    #   supabase-auth
pydantic-core==2.41.5
    # via pydantic
pydantic-settings==2.12.0
    # via mcp
pydub==0.25.1
    # via gradio
pygments==2.19.2
    # via
    #   pytest
    #   rich
pyiceberg==0.10.0
    # via storage3
pyjwt==2.10.1
    # via
    #   mcp
    #   supabase-auth
pyparsing==3.3.2
    # via pyiceberg
pyroaring==1.0.3
    # via pyiceberg
pytest==9.0.2
    # via
    #   ecommerce-support
    #   pytest-asyncio
    #   pytest-cov
pytest-asyncio==1.3.0
    # via ecommerce-support
pytest-cov==7.0.0
    # via ecommerce-support
python-dateutil==2.9.0.post0
    # via
    #   pandas
    #   strictyaml
python-dotenv==1.2.1
    # via pydantic-settings
python-multipart==0.0.21
    # via
    #   gradio
    #   mcp
pytz==2025.2
    # via pandas
pywin32==311 ; sys_platform == 'win32'
    # via mcp
pyyaml==6.0.3
    # via
    #   gradio
    #   huggingface-hub
realtime==2.27.2
    # via supabase
referencing==0.37.0
    # via
    #   jsonschema
    #   jsonschema-specifications
requests==2.32.5
    # via
    #   openai-agents
    #   pyiceberg
rich==14.2.0
    # via
    #   pyiceberg
    #   typer
rpds-py==0.30.0
    # via
    #   jsonschema
    #   referencing
safehttpx==0.1.7
    # via gradio
semantic-version==2.10.0
    # via gradio
shellingham==1.5.4
    # via
    #   huggingface-hub
    #   typer
six==1.17.0
    # via python-dateutil
sniffio==1.3.1
    # via openai
sortedcontainers==2.4.0
    # via pyiceberg
sse-starlette==3.2.0
    # via mcp
starlette==0.50.0
    # via
    #   fastapi
    #   gradio
    #   mcp
    #   sse-starlette
storage3==2.27.2
    # via supabase
strenum==0.4.15
    # via supabase-functions
strictyaml==1.7.3
    # via pyiceberg
supabase==2.27.2
    # via ecommerce-support
supabase-auth==2.27.2
    # via supabase
supabase-functions==2.27.2
    # via supabase
tenacity==9.1.2
    # via pyiceberg
tomlkit==0.13.3
    # via gradio
tqdm==4.67.1
    # via
    #   huggingface-hub
    #   openai
typer==0.21.1
    # via gradio
typer-slim==0.21.1
    # via huggingface-hub
types-requests==2.32.4.20260107
    # via openai-agents
typing-extensions==4.15.0
    # via
    #   fastapi
    #   gradio
    #   gradio-client
    #   huggingface-hub
    #   mcp
    #   openai
    #   openai-agents
    #   pydantic
    #   pydantic-core
    #   realtime
    #   typer
    #   typer-slim
    #   typing-inspection
typing-inspection==0.4.2
    # via
    #   mcp
    #   pydantic
    #   pydantic-settings
tzdata==2025.3
    # via pandas
urllib3==2.6.3
    # via
    #   requests
    #   types-requests
uvicorn==0.40.0
    # via
    #   gradio
    #   mcp
websockets==15.0.1
    # via realtime
yarl==1.22.0
    # via
    #   postgrest
    #   storage3
    #   supabase
    #   supabase-functions
